Add files via upload
Optimization of the angles of solar panels and location-allocation models using linear programming solved with LINGO
The Lingo files for the main issues are OR-Q1 and OR-Q2, respectively; the files for the sensitivity analysis of the right-hand side constraints are RHS and Q2-RHS, respectively, and the files related to the sensitivity analysis of the variable coefficient are tabe hadaf and Q2-tabe hadaf, respectively. The PDF with explanations is also included.